server/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
djangorestframework==3.15.2
joblib==1.4.2
kombu==5.4.2
numpy==2.1.3
pandas==2.2.3
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
sqlparse==0.5.2
threadpoolctl==3.5.0
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
